Can finally render objects
There are still some problems with the object renderer that I need to
fix. The first one being that for some reason there is a scaling issue
with non-full size textures, for things like the corners where there are
only a few pixels but the whole thing is rendered to the destination
rect which is fixed. The other is that there is still an offset that i'm
missing with some of the objects, as they don't appear where the should
be relative to the tile they are supposed to be rendered on.