Add "wire-tapping" server
This program will to listen in and decode packets sent between
the client and server. It's currently configured to catch traffic from
the client which is in a virtual machine and a modified `info.dll` which
points the client to the server program. Once a connection is made from
the client the local server will open a connection with the _actual_
server and just forward the packets back and forth.

It seems however that there is a handoff of the server to another
address or port after the actual login session is finished (after the
version check)